agregando alert confirmacion y cancelacion

+++ proyectoFinalApp/src/pages/CarritoPage.jsx
@@ -1,10 +1,26 @@
+import { dispararSweetBasico } from "../assets/SweetAlert";
 import Carrito from "../components/Carrito";
 
 function CarritoPage({productosCarrito, eliminarDelCarrito}) {
 
     function eliminarDelCarritoPages(productoAEliminar) {
-        eliminarDelCarrito(productoAEliminar);
+        dispararSweetBasico(
+            "¿Estás seguro de eliminar este producto del carrito?",
+            "Esta acción no se puede deshacer.",
+            "warning",
+            "Sí, eliminar",
+            true
+        ).then((result) => {
+            console.log(result);
+            
+            if (result.isConfirmed) {
+                eliminarDelCarrito(productoAEliminar);
+            }
+        });
+        
+        // eliminarDelCarrito(productoAEliminar);
     }
+
     
     return(
         <Carrito productos={productosCarrito} eliminarDelCarritoPages={eliminarDelCarritoPages} />
